% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcharts.R
\name{add_time_legend}
\alias{add_time_legend}
\title{Add a text-to-speech legend for the time axis.}
\usage{
add_time_legend(x, break_values = pretty(x, 3), chart_length)
}
\arguments{
\item{x}{@break_values Time values for the whole plot and the legend respectively.}
}
\description{
Add a text-to-speech legend for the time axis.
}
