% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcharts.R
\name{compose_dotplot}
\alias{compose_dotplot}
\title{Make a dotplot where x axis is time, y axis is pitch, and each dot is a beep.}
\usage{
compose_dotplot(
  data,
  main,
  time,
  pitch,
  filename,
  handle_collisions = "superimpose",
  envelope.shape = "hairpin",
  chart.duration = 10,
  beep.duration = 0.1,
  ...
)
}
\arguments{
\item{data}{Dataframe in long format.}

\item{main}{Character of length 1. Chart title.}

\item{time}{@param pitch Columns from \code{data}.}

\item{filename}{Where to put results. This should end in \code{.wav}.}

\item{chart.duration}{How many seconds the sound file plays for.}

\item{beep.duration}{How many seconds an individual beep or point lasts.}

\item{...}{Args sent to scale_pitch().}
}
\description{
Make a dotplot where x axis is time, y axis is pitch, and each dot is a beep.
}
