% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcharts.R
\name{get_spoken_words}
\alias{get_spoken_words}
\title{Given text, return a tuneR wav file object where said text is ... said.}
\usage{
get_spoken_words(words)
}
\description{
Given text, return a tuneR wav file object where said text is ... said.
}
