% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcharts.R
\name{scale_pitch}
\alias{scale_pitch}
\title{Convert an abstract set of numbers into frequencies suitable for human ears.}
\usage{
scale_pitch(x, lower = 220, upper = 880, relative = T)
}
\arguments{
\item{lower}{@param upper Most extreme pitches.
Default is A440 plus or minus an octave.}

\item{relative}{If false, equal intervals in data are separated by equal
intervals in absolute pitch (equal numbers of Hz).
If true, equal intervals in data are separated by equal
intervals in relative pitch (equal numbers of cents or semitones or octaves).}
}
\description{
Convert an abstract set of numbers into frequencies suitable for human ears.
}
