% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sp500}
\alias{sp500}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 6624 rows and 4 variables:
\describe{
\item{Close}{Share price, in US dollars}
\item{Date}{Dates in original format}
\item{DateLubridate}{Dates in the format used by the lubridate package}
\item{DaysSinceStart}{Days since the start of the time-series}
...
}
}
\source{
\url{https://www.investopedia.com/ask/answers/042415/what-average-annual-return-sp-500.asp}
}
\usage{
sp500
}
\description{
A dataset containing s&p 500 performance since 1993.
}
\keyword{datasets}
